top 10 function, seperate functionality
